ITI full stack python/python/crowdfunding_app/requirements.txt
psycopg2-binary
# bcrypt
